+++ NEWS.md
@@ -19,10 +19,17 @@
   sessions.
 - Add `predict` method to obtain conditional mean estimates and prediction intervals, both with or without previous observations.
 - Use automatic differentiation to calculate Satterthwaite adjusted degrees of freedom calculation, resulting in 10-fold speed-up of the Satterthwaite calculations after the initial model fit.
+- Add methods `model.matrix()` and `terms()` often available for modeling functions that assist in post-processing of a model object.
 
 ### Miscellaneous
 
 - Removed `free_cores()` in favor of `parallelly::availableCores(omit = 1)`.
+- The `model.frame()` method has been updated to return a data frame the size of 
+  the number of observations utilized in the model for all combinations of the 
+  `model.frame(include)` argument when `na.action='na.omit'` (the default).
+- The `model.frame(include=)` method argument's default has been updated 
+  from `include=NULL` to `include=c("subject_var", "visit_var", "group_var", "response_var")`
+  such that by default all relevant variables are returned.
 
 ### Bug Fixes